Refactored project definitions + how they are saved
Projects are now always saved as soon as they are created, meaning there is no extra work required with determining if a project has a saved path when they are being used. This is useful because we require this path for relative paths to be possible in the new tasks and in general when it comes to saving these paths.